add SAC and PPO configuration files; remove obsolete config and main.py
